refactor: add auth layout
