+++ src/components/common/index.ts
@@ -2,4 +2,3 @@ export * from './AuthButton';
 export * from './AuthInput';
 export * from './Balloon';
 export * from './Fallback';
-export * from './ClientComponentLayout';
